fix(vite): improve code spliting
close #6202
