--- packages/vite/src/dirs.ts
@@ -1,5 +1,0 @@
-import { fileURLToPath } from 'node:url'
-import { dirname, resolve } from 'pathe'
-
-export const distDir = dirname(fileURLToPath(import.meta.url))
-export const pkgDir = resolve(distDir, '..')
