+++ packages/vite/src/utils/dirs.ts
@@ -1,0 +1,5 @@
+import { fileURLToPath } from 'node:url'
+import { resolve } from 'pathe'
+
+export const distDir = resolve(fileURLToPath(import.meta.url), '../..')
+export const pkgDir = resolve(distDir, '..')
