+++ packages/vite/src/vite-node.ts
@@ -6,7 +6,7 @@ import { resolve } from 'pathe'
 import { addServerMiddleware } from '@nuxt/kit'
 import type { Plugin as VitePlugin, ViteDevServer } from 'vite'
 import { resolve as resolveModule } from 'mlly'
-import { distDir } from './dirs'
+import { distDir } from './utils/dirs'
 import type { ViteBuildContext } from './vite'
 import { isCSS } from './utils'
 import { createIsExternal } from './utils/external'
